+++ CHANGELOG.md
@@ -2,6 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Commodum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Commodum/releases)
 
-- Added: General loot notifications for items, money, and personal currencies can now be hidden
-- Minor code adjustments
-- Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility
+- Added: Damaged items can now be repaired automatically with personal funds when a repair merchant window opens